pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocr-web"
classifiers = ["Private :: Do Not Upload"]
version = "0"
requires-python = ">=3.9"
dependencies = [
    "Flask[async]",
    "gunicorn",
    "Jinja2",
    "pillow",
    "pytesseract",
    "werkzeug",
    "pdf2image",
    "langcodes[data]==3.5.1",
    "language-data==1.4.0",
    "requests",
    "pandas",
    "mcp",
]


[project.optional-dependencies]
dev = [
    "pip-tools",
    "isort",
    "black",
    "ruff",
    "tox",
    "pytest",
    "pytest-cov",
]


[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
